feat : comment RUD 검증 추가

+++ BE/Blog/src/main/java/hyundai/blog/comment/service/CommentService.java
@@ -3,7 +3,13 @@ package hyundai.blog.comment.service;
 import hyundai.blog.comment.dto.CommentCreateRequest;
 import hyundai.blog.comment.dto.CommentUpdateRequest;
 import hyundai.blog.comment.entity.Comment;
+import hyundai.blog.comment.exception.CommentIdNotFoundException;
+import hyundai.blog.comment.exception.InvalidCommentOwnerException;
 import hyundai.blog.comment.repository.CommentRepository;
+import hyundai.blog.member.entity.Member;
+import hyundai.blog.til.entity.Til;
+import hyundai.blog.til.exception.InvalidTilOwnerException;
+import hyundai.blog.util.MemberResolver;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.stereotype.Service;
@@ -15,11 +21,16 @@ import java.time.LocalDateTime;
 @Log4j2
 public class CommentService {
     private final CommentRepository commentRepository;
+    private final MemberResolver memberResolver;
+
 
     public Comment save(CommentCreateRequest request) {
+        // 현재 로그인 된 멤버의 ID를 가져온다.
+        Member loggedInMember = memberResolver.getCurrentMember();
+
         // 1) dto 를 가지고 entity를 만들어주고
         Comment comment = Comment.builder()
-                .memberId(request.getMemberId())
+                .memberId(loggedInMember.getId())
                 .tilId(request.getTilId())
                 .content(request.getContent())
                 .createdAt(LocalDateTime.now())
@@ -35,7 +46,13 @@ public class CommentService {
     }
 
     public Comment update(Long id, CommentUpdateRequest request) {
-        Comment comment = commentRepository.findById(id).orElseThrow(IllegalArgumentException::new);
+        // 현재 로그인 된 멤버의 ID를 가져온다.
+        Member loggedInMember = memberResolver.getCurrentMember();
+
+        Comment comment = commentRepository.findById(id).orElseThrow(CommentIdNotFoundException::new);
+
+        // 3) til 작성자와 로그인한 사용자에 대한 검증
+        validateCommentOwnerShip(comment, loggedInMember);
 
         comment.changeContent(request.getContent());
         comment.changeUpdatedAt(LocalDateTime.now());
@@ -46,6 +63,19 @@ public class CommentService {
     }
 
     public void delete(Long id) {
+
+        Member loggedInMember = memberResolver.getCurrentMember();
+
+        Comment comment = commentRepository.findById(id).orElseThrow(CommentIdNotFoundException::new);
+
+        validateCommentOwnerShip(comment, loggedInMember);
+
         commentRepository.deleteById(id);
+    }
+
+    private void validateCommentOwnerShip(Comment comment, Member loggedInMember) {
+        if (!comment.getMemberId().equals(loggedInMember.getId())) {
+            throw new InvalidCommentOwnerException();
+        }
     }
 }
